Split up ESP32 and ESP8266 wifi logic
